requirement.txt
absl-py==1.4.0
annotated-types==0.7.0
anyio==4.4.0
apeye==1.3.0
apeye-core==1.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.0.7
cycler==0.11.0
DateTime==5.5
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.1
djangorestframework==3.14.0
dm-tree==0.1.8
dnspython==2.6.1
domdf-python-tools==3.6.1
EasyProcess==1.1
email_validator==2.2.0
entrypoint2==1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.112.0
fastapi-cli==0.0.5
fastjsonschema==2.19.1
filelock==3.13.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
gast==0.4.0
geopandas==1.0.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.32.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.0
ipympl==0.9.3
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keras==2.13.1
keras-core==0.1.5
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
moviepy==1.0.3
mss==9.0.1
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.4
namex==0.0.7
natsort==8.3.1
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
newsapi-python==0.2.7
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
osmnx==1.9.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.3
Pillow==9.5.0
platformdirs==4.1.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.24.3
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
PyMsgBox==1.0.9
pyogrio==0.9.0
pyparsing==3.0.9
pyperclip==1.8.2
pypi-json==0.3.0
pypng==0.20220715.0
pyproj==3.6.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytube==15.0.0
pytube3==9.6.4
pytweening==1.0.7
pytz==2024.1
pywhatkit==5.4
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.2
pyzmq==25.1.2
qrcode==7.4.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.10.1
seaborn==0.12.2
selenium==4.16.0
Send2Trash==1.8.2
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.30.5
virtualenv==20.25.0
watchfiles==0.23.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.3.4
widgetsnbextension==4.0.10
wikipedia==1.4.0
wrapt==1.15.0
wsproto==1.2.0
zope.interface==7.0.1
